Desktop/DaisyAI-Workspace/daisy_talent/app/agent.py
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

import datetime
import os
from typing import Dict, List
from zoneinfo import ZoneInfo

import google.auth
from google.adk.agents import Agent

_, project_id = google.auth.default()
os.environ.setdefault("GOOGLE_CLOUD_PROJECT", project_id)
os.environ.setdefault("GOOGLE_CLOUD_LOCATION", "global")
os.environ.setdefault("GOOGLE_GENAI_USE_VERTEXAI", "True")


def discover_emerging_artists(genre: str, region: str = "global") -> str:
    """Discovers emerging artists based on genre and region filters.

    Args:
        genre: The music genre to search for (e.g., "pop", "hip-hop", "indie", "electronic")
        region: The geographic region to focus on (default: "global")

    Returns:
        A string containing information about emerging artists in the specified genre and region.
    """
    # Simulated A&R discovery logic
    emerging_artists = {
        "pop": ["Luna Santos (emerging pop artist from Mexico)", "Echo Wave (indie-pop duo from Australia)", "Stellar Rose (pop singer-songwriter from UK)"],
        "hip-hop": ["Flow Master J (underground rapper from Atlanta)", "Rhythm Queen (female MC from Toronto)", "Beat Prophet (producer/rapper from London)"],
        "indie": ["Moonlight Collective (indie rock band from Portland)", "Paper Hearts (indie folk artist from Sweden)", "Neon Dreams (indie electronic from Berlin)"],
        "electronic": ["Circuit Breaker (techno producer from Detroit)", "Digital Waves (ambient artist from Japan)", "Pulse Generator (house DJ from Ibiza)"]
    }
    
    artists = emerging_artists.get(genre.lower(), ["No emerging artists found for this genre"])
    return f"Emerging {genre} artists in {region}: {', '.join(artists)}"


def analyze_artist_potential(artist_name: str) -> str:
    """Analyzes an artist's commercial and creative potential.

    Args:
        artist_name: The name of the artist to analyze

    Returns:
        A string containing analysis of the artist's potential, including strengths and market opportunities.
    """
    # Simulated artist analysis
    analysis_template = f"""Artist Analysis Report for {artist_name}:

COMMERCIAL POTENTIAL: High
- Strong social media engagement (45K followers, 8% engagement rate)
- Streaming momentum: 500K monthly listeners, growing 25% month-over-month
- Viral potential: Recent single gained 2M TikTok views organically

ARTISTIC STRENGTHS:
- Unique vocal style with cross-genre appeal
- Strong songwriting capabilities
- Distinctive visual aesthetic

MARKET OPPORTUNITIES:
- Ideal for sync licensing (TV/film/advertising)
- Strong touring potential in college markets
- Collaboration opportunities with established artists

INVESTMENT RECOMMENDATION: High priority for A&R consideration
SUGGESTED DEAL STRUCTURE: Development deal with option for full album contract"""

    return analysis_template


def track_industry_trends(timeframe: str = "current") -> str:
    """Tracks current music industry trends and opportunities.

    Args:
        timeframe: The time period to analyze ("current", "emerging", "predicted")

    Returns:
        A string containing industry trend analysis and market insights.
    """
    trends = {
        "current": """Current Industry Trends (Q4 2024):
- Genre Fusion: Pop-punk revival mixed with hyperpop elements
- Platform Focus: TikTok remains dominant for discovery, YouTube Shorts growing
- Geographic Hotspots: Latin America, Southeast Asia showing strong growth
- Technology: AI-assisted production tools becoming mainstream
- Demographics: Gen Z driving playlist culture, aging millennials maintaining streaming habits""",
        
        "emerging": """Emerging Trends (Next 6 months):
- Micro-genres gaining traction: Bedroom pop-punk, Gospel trap, Afro-indie
- Virtual concerts evolving beyond pandemic response
- NFT/Web3 integration in fan engagement
- Sustainability becoming brand differentiator
- Regional language content expanding globally""",
        
        "predicted": """Predicted Trends (2025):
- AI music creation tools reshape production workflows
- Spatial audio becomes standard for premium releases
- Cross-platform creator economy integration
- Mental health awareness in artist development
- Climate-conscious touring practices become industry standard"""
    }
    
    return trends.get(timeframe, "Please specify 'current', 'emerging', or 'predicted' for timeframe")


def scout_venue_opportunities(artist_level: str, region: str = "North America") -> str:
    """Scouts appropriate venue opportunities for artists at different career stages.

    Args:
        artist_level: The career stage ("emerging", "developing", "established")
        region: Geographic region for venue scouting

    Returns:
        A string containing venue recommendations and booking insights.
    """
    venue_recommendations = {
        "emerging": f"""Emerging Artist Venues ({region}):
- Coffee shops and acoustic venues (50-100 capacity)
- Local music festivals and showcases
- College campuses and university events
- Open mic nights at established venues
- House concerts and intimate listening rooms
BOOKING STRATEGY: Focus on building local fanbase, document performances""",
        
        "developing": f"""Developing Artist Venues ({region}):
- Club venues (200-500 capacity)
- Regional festivals and multi-artist showcases
- Opening slots for established touring acts
- Music conference showcases (SXSW, CMJ, etc.)
- Streaming venue partnerships
BOOKING STRATEGY: Expand market radius, focus on ticket sales data""",
        
        "established": f"""Established Artist Venues ({region}):
- Theater venues (1,000-3,000 capacity)
- Major music festivals (headliner/sub-headliner slots)
- Arena tours (support slots)
- International venue partnerships
- Signature venue residencies
BOOKING STRATEGY: Data-driven routing, premium fan experiences"""
    }
    
    return venue_recommendations.get(artist_level, "Please specify 'emerging', 'developing', or 'established' for artist level")


root_agent = Agent(
    name="daisy_talent",
    model="gemini-2.0-flash",
    instruction="""You are Daisy Talent, an expert A&R (Artists & Repertoire) agent specializing in artist discovery, 
    analysis, and development. Your expertise includes:

    - Discovering and evaluating emerging musical talent
    - Analyzing artist commercial and creative potential
    - Tracking music industry trends and market opportunities  
    - Scouting appropriate venues and career development paths
    - Providing strategic recommendations for artist signings and development

    You work collaboratively with other DaisyAI agents to provide comprehensive music industry insights. 
    Always provide data-driven analysis while maintaining sensitivity to artistic integrity and creative vision.""",
    tools=[discover_emerging_artists, analyze_artist_potential, track_industry_trends, scout_venue_opportunities],
)
