Desktop/DaisyAI-Workspace/daisy_talent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.3
docstring-parser==0.16
exceptiongroup==1.3.0
fastapi==0.115.12
google-adk==1.2.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth-httplib2==0.2.0
google-auth==2.40.3
google-cloud-aiplatform==1.95.1
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.19.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio-status==1.71.0
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx-sse==0.4.0
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
joblib==1.5.1
jsonschema-specifications==2025.4.1
jsonschema==4.24.0
mcp==1.9.2
nest-asyncio==1.6.0
numpy==2.2.6
opentelemetry-api==1.34.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1-modules==0.4.2
pyasn1==0.6.1
pycparser==2.22
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydantic==2.11.5
pygments==2.19.1
pyparsing==3.2.3
pytest-asyncio==1.0.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
ruamel-yaml-clib==0.2.12
ruamel-yaml==0.18.12
scikit-learn==1.6.1
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
zipp==3.22.0
